Add or move unit tests for files in link folder
